refactor(LottoGame): 결과 출력 화면 로직 분리- ResultPrinter
